+++ frontend/src/components/question-creation/QuestionPage.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
-
 import { Flex } from "@chakra-ui/react";
-import QuestionSidebar from "./QuestionSidebar";
+
 import { HOME_PAGE } from "../../constants/Routes";
+import QuestionSidebar from "./QuestionSidebar";
 import QuestionEditor from "./QuestionEditor";
 
-import QuestionElement from "./types/QuestionTypes";
+import { QuestionElement } from "../../types/QuestionTypes";
 
 const QuestionPage = (): React.ReactElement => {
   const [questionElements, setQuestionElements] = React.useState<
