--- frontend/src/components/question-creation/types/QuestionTypes.tsx
@@ -1,10 +1,0 @@
-enum QuestionElement {
-  QUESTION = "Question",
-  TEXT = "Text",
-  IMAGE = "Image",
-  MULTIPLE_CHOICE = "Multiple Choice",
-  SHORT_ANSWER = "Short Answer",
-  MULTI_SELECT = "Multi-select",
-}
-
-export default QuestionElement;
