+++ frontend/src/types/QuestionTypes.ts
@@ -1,7 +1,14 @@
-enum QuestionType {
+export enum QuestionType {
   MULTIPLE_CHOICE = "Multiple Choice",
   SHORT_ANSWER = "Short Answer",
   MULTI_SELECT = "Multi-select",
 }
 
-export default QuestionType;
+export enum QuestionElement {
+  QUESTION = "Question",
+  TEXT = "Text",
+  IMAGE = "Image",
+  MULTIPLE_CHOICE = "Multiple Choice",
+  SHORT_ANSWER = "Short Answer",
+  MULTI_SELECT = "Multi-select",
+}
